refactor: rename analysis workflow provider module
