+++ tests/test_analysis_workflow_port.py
@@ -3,7 +3,7 @@ import unittest
 from tests import _path  # noqa: F401
 
 from qfit.analysis.application.analysis_controller import AnalysisController
-from qfit.analysis.application.analysis_controller_provider import build_analysis_workflow
+from qfit.analysis.application.analysis_workflow_provider import build_analysis_workflow
 from qfit.analysis.application.analysis_workflow_port import AnalysisWorkflowPort
